+++ src/freenet/node/fcp/ProbeLinkLengths.java
@@ -6,7 +6,7 @@ import freenet.io.comm.DMT;
  * FCP message sent from the node to the client which includes link lengths reported by the endpoint.
  */
 public class ProbeLinkLengths extends FCPResponse {
-	public ProbeLinkLengths(String fcpIdentifier, double[] linkLengths) {
+	public ProbeLinkLengths(String fcpIdentifier, float[] linkLengths) {
 		super(fcpIdentifier);
 		fs.putOverwrite(FCPMessage.IDENTIFIER, fcpIdentifier);
 		fs.put(DMT.LINK_LENGTHS, linkLengths);
